Move most of integ tests over to unit
