+++ test/integ/handlers/on-chain-status-checker.integration.test.ts
@@ -10,8 +10,8 @@ import { getProvider, getValidator, getWatcher } from '../../../lib/handlers/che
 import { log } from '../../../lib/Logging'
 import { OnChainStatusCheckerMetricNames, powertoolsMetric } from '../../../lib/Metrics'
 import { LimitOrdersRepository } from '../../../lib/repositories/limit-orders-repository'
+import { dynamoConfig, MOCK_ORDER_ENTITY, MOCK_ORDER_HASH, MOCK_SIGNATURE } from '../../test-data'
 import { deleteAllRepoEntries } from '../../unit/utils'
-import { dynamoConfig, MOCK_ORDER_ENTITY, MOCK_ORDER_HASH, MOCK_SIGNATURE } from './test-data'
 
 jest.mock('../../../lib/handlers/check-order-status/util', () => {
   const original = jest.requireActual('../../../lib/handlers/check-order-status/util')
