+++ test/test-data.ts
@@ -1,6 +1,5 @@
-/* eslint-disable */
 import { OrderType, REACTOR_ADDRESS_MAPPING } from '@uniswap/uniswapx-sdk'
-import { OrderEntity, ORDER_STATUS } from '../../../lib/entities'
+import { OrderEntity, ORDER_STATUS } from '../lib/entities'
 
 export const MOCK_ORDER_HASH = '0xc57af022b96e1cb0da0267c15f1d45cdfccf57cfeb8b33869bb50d7f478ab203'
 export const MOCK_ENCODED_ORDER =
